pyproject.toml
[tool.poetry]
name = "alfa-photo-share"
version = "0.1.0"
description = ""
authors = ["alfa team <alfa-team@alfa.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
psycopg2-binary = "^2.9.10"
alembic = "^1.15.2"
python-jose = "^3.4.0"
redis = "^5.2.1"
passlib = "^1.7.4"
bcrypt = "4.0.1"
pydantic = {extras = ["email"], version = "^2.11.3"}
fastapi-limiter = "^0.1.6"
pydantic-settings = "^2.9.1"
asyncpg = "^0.30.0"
python-multipart = "^0.0.20"
cloudinary = "^1.44.0"
qrcode = "^8.1"
pillow = "^11.2.1"
httpx = "^0.28.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
httpx = "^0.28.1"
pytest-asyncio = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
